requirements.txt
numpy==1.14.5
scipy>=1.0.0
six>=1.10.0
tensorflow-gpu==1.14.0
